job #10195 added two logical checks for qualifying type references
First, check if the type is based on 'MASLtype'. If it is, this is a
type reference type and should never be qualified with the containing
domain name.

Second, check if the package is directly off the system. In this case,
it is assumed that the user understands that the package is not contained
by a component so it will _not_ be qualified. A special case that is
important for this check is the 'types' package which contains MASL core
types and is generated directly off the system. This check prevents types
in this package from ever being qualified.